fixed incorrect type reference in codon_counts() docstrings
int -> float for threshold arguments

+++ src/vhip/mlmodel/gene_features.py
@@ -192,8 +192,8 @@ class GeneSet:
         """Aggregate the counts for each unique codon and imprecise codons across an entire GeneSet.
 
         Args:
-            threshold_imprecise (int): Percentage of imprecise (non-ATGC) codons tolerated in a single gene (default 0.0 or 0%)
-            threshold_skipped_genes (int): Tolerated percentage of valid (codon length divisible) genes in GeneSet that have more than threshold_imprecise codons (default 0.5 or 50%)
+            threshold_imprecise (float): Percentage of imprecise (non-ATGC) codons tolerated in a single gene (default 0.0 or 0%)
+            threshold_skipped_genes (float): Tolerated percentage of valid (codon length divisible) genes in GeneSet that have more than threshold_imprecise codons (default 0.5 or 50%)
         Populates the following class attributes:
             self.codon_dict (str: int): Counts of each unique codon across all genes in the GeneSet.
             self.imprecise_codons (int): Total number of imprecise codons found in the GeneSet.
